git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/addons-fwork@26324 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/cmyth/Win32/include/mysql/my_alloc.h
@@ -37,8 +37,8 @@ typedef struct st_mem_root
   USED_MEM *used;                  /* blocks almost without free memory */
   USED_MEM *pre_alloc;             /* preallocated block */
   /* if block have less memory it will be put in 'used' list */
-  unsigned int min_malloc;
-  unsigned int block_size;         /* initial block size */
+  size_t min_malloc;
+  size_t block_size;               /* initial block size */
   unsigned int block_num;          /* allocated blocks counter */
   /* 
      first free block in queue test counter (if it exceed
